backend/requirements.txt
torch==2.0.1
transformers==4.28.0
spacy==3.5.2
scikit-learn==1.2.2
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0   # if you want to load .env automatically
openai==0.27.2
tqdm==4.65.0
pandas==2.0.0
pydantic==1.10.7
# ...any other libs you used
